Add OVS tools
This adds a python script and a shell script, which can be used to
recreate a dataplane (br-int and br-fabric ports, along with flow-mods).
The bridge_ifs.py python module is used to create a script file (default
is config-bridges.sh) that creates and adds tap ports, SNAT ports, and
patch ports to the br-int and br-fabric bridges. The create-flowmods.sh
shell script is used to sanitize the output of a ovs-ofctl dump-flows
command, and converts that into individual ovs-ofctl add-flow commands
in a file, which can be run to create the same flow-mods in OVS.