New commands and fix
This adds commands to count the number of MOs, as well as a fix for
policy comparisons. This change relaces the compare for the order
of children declared by the parent. As long as the child MOs are
the same, their order shouldn't matter when doing logical policy
comparisons. It also adds the ability to handle AIM hash tree
dumps.